Upstream from froge

+++ src/main/java/com/simibubi/create/content/contraptions/fluids/FluidNetwork.java
@@ -183,7 +183,7 @@ public class FluidNetwork {
 			});
 		}
 
-		int flowSpeed = transferSpeed;
+		long flowSpeed = transferSpeed;
 		for (boolean simulate : Iterate.trueAndFalse) {
 
 			IFluidHandler handler = source.orElse(null);
